move files to correct package dir
